+++ recipes/TensorFlow-Distributed/docker/cpu/launcher.sh
@@ -1,5 +1,8 @@
 #!/usr/bin/env bash
 
+set -e
+set -o pipefail
+
 # get my ip address
 ipaddress=`ip addr list eth0 | grep "inet " | cut -d' ' -f6 | cut -d/ -f1`
 
@@ -32,12 +35,17 @@ echo "ps hosts: $ps_hosts"
 echo "worker hosts: $worker_hosts"
 
 # master node acts as parameter server
+masterpid=
 if [ $AZ_BATCH_IS_CURRENT_NODE_MASTER == "true" ]; then
     # master node
     ti=${task_index[$master]}
     echo "master node: $ipaddress task index: $ti"
     python /sw/mnist_replica.py --ps_hosts=$ps_hosts --worker_hosts=$worker_hosts --job_name=ps --task_index=$ti --data_dir=./master --num_gpus=0 $* > ps-$ti.log 2>&1 &
+    masterpid=$!
 fi
+
+declare -a waitpids
+
 # launch worker nodes
 for node in "${HOSTS[@]}"
 do
@@ -45,7 +53,30 @@ do
     echo "worker node: $node task index: $ti"
     if [ $node == $master ]; then
         python /sw/mnist_replica.py --ps_hosts=$ps_hosts --worker_hosts=$worker_hosts --job_name=worker --task_index=$ti --data_dir=./worker-$ti --num_gpus=0 $* > worker-$ti.log 2>&1 &
+        waitpids=("${waitpids[@]}" "$!")
     else
-        ssh $node "python /sw/mnist_replica.py --ps_hosts=$ps_hosts --worker_hosts=$worker_hosts --job_name=worker --task_index=$ti --data_dir=$AZ_BATCH_TASK_WORKING_DIR/worker-$ti --num_gpus=0 $* > $AZ_BATCH_TASK_WORKING_DIR/worker-$ti.log 2>&1 &"
+        ssh $node "python /sw/mnist_replica.py --ps_hosts=$ps_hosts --worker_hosts=$worker_hosts --job_name=worker --task_index=$ti --data_dir=$AZ_BATCH_TASK_WORKING_DIR/worker-$ti --num_gpus=0 $* > $AZ_BATCH_TASK_WORKING_DIR/worker-$ti.log 2>&1" &
+        waitpids=("${waitpids[@]}" "$!")
     fi
 done
+
+# because the grpc server does not automatically exit, we need to
+# wait for all of the child processes in waitpids to complete first
+declare -a donepids
+while :
+do
+    for pid in "${waitpids[@]}"; do
+        kill -0 $pid
+        if [ $? -ne 0 ]; then
+            donepids=("${donepids[@]}" "$pid")
+        fi
+    done
+    if [ ${#waitpids[@]} -eq ${#donepids[@]} ]; then
+        break
+    else
+        sleep 1
+    fi
+done
+
+# kill master process
+kill -9 $masterpid